Refator : dto 기능분할
