+++ backend/src/main/java/kr/co/programmers/collabond/api/recruit/interfaces/RecruitPostController.kt
@@ -3,8 +3,8 @@ package kr.co.programmers.collabond.api.recruit.interfaces
 import jakarta.validation.Valid
 import kr.co.programmers.collabond.api.recruit.application.RecruitPostService
 import kr.co.programmers.collabond.api.recruit.domain.RecruitPostStatus
-import kr.co.programmers.collabond.api.recruit.domain.dto.RecruitPostRequestDto
-import kr.co.programmers.collabond.api.recruit.domain.dto.RecruitPostResponseDto
+import kr.co.programmers.collabond.api.recruit.domain.dto.Requests
+import kr.co.programmers.collabond.api.recruit.domain.dto.Responses
 import kr.co.programmers.collabond.core.auth.oauth2.OAuth2UserInfo
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -22,22 +22,22 @@ class RecruitPostController(
 
     @PostMapping
     fun createRecruitPost(
-        @Valid @RequestBody request: RecruitPostRequestDto,
+        @Valid @RequestBody request: Requests,
         @AuthenticationPrincipal userInfo: OAuth2UserInfo
-    ): ResponseEntity<RecruitPostResponseDto> =
+    ): ResponseEntity<Responses> =
         ResponseEntity.status(HttpStatus.CREATED)
             .body(recruitPostService.createRecruitPost(request, userInfo))
 
     @GetMapping("/{recruitmentId}")
-    fun getRecruitPostById(@PathVariable recruitmentId: Long): ResponseEntity<RecruitPostResponseDto> =
+    fun getRecruitPostById(@PathVariable recruitmentId: Long): ResponseEntity<Responses> =
         ResponseEntity.ok(recruitPostService.getRecruitPostById(recruitmentId))
 
     @PatchMapping("/{recruitmentId}")
     fun updateRecruitPost(
         @PathVariable recruitmentId: Long,
-        @Valid @RequestBody request: RecruitPostRequestDto,
+        @Valid @RequestBody request: Requests,
         @AuthenticationPrincipal userInfo: OAuth2UserInfo
-    ): ResponseEntity<RecruitPostResponseDto> =
+    ): ResponseEntity<Responses> =
         ResponseEntity.ok(recruitPostService.updateRecruitPost(recruitmentId, request, userInfo))
 
     @DeleteMapping("/{recruitmentId}")
@@ -54,20 +54,20 @@ class RecruitPostController(
         @RequestParam(required = false) status: RecruitPostStatus? = null,
         @RequestParam(required = false) sort: String? = null,
         @PageableDefault(size = 10) pageable: Pageable
-    ): ResponseEntity<Page<RecruitPostResponseDto>> =
+    ): ResponseEntity<Page<Responses>> =
         ResponseEntity.ok(recruitPostService.getAllRecruitPosts(status, sort, pageable))
 
     @GetMapping("/users/{userId}")
     fun getRecruitPostsByUser(
         @PathVariable userId: Long,
         @PageableDefault(size = 10) pageable: Pageable
-    ): ResponseEntity<Page<RecruitPostResponseDto>> =
+    ): ResponseEntity<Page<Responses>> =
         ResponseEntity.ok(recruitPostService.getRecruitPostsByUser(userId, pageable))
 
     @GetMapping("/profiles/{profileId}")
     fun getRecruitPostByProfile(
         @PathVariable profileId: Long,
         @PageableDefault(size = 10) pageable: Pageable
-    ): ResponseEntity<Page<RecruitPostResponseDto>> =
+    ): ResponseEntity<Page<Responses>> =
         ResponseEntity.ok(recruitPostService.getRecruitPostByProfile(profileId, pageable))
 }
